ENT-2769: Speed up local and remote file copying and fix spurious errors.
Changelog: Title

When writing to disk files, zeroes are automatically being detected and
sparse files are being created to save disk space. This was happening
even for small "holes" of a couple of zeroes, and the system call
overhead was high.

It was fixed by writing "holes", i.e. making the file sparse, only if
the holes of zeroes were longer than the block size.

Also fixes a seemingly spurious failure to write files (ftruncate() was
returning EPERM). Reason is that on files that ended with zeroes, we
used to call ftruncate() with an offset past the end of the file. This
is not guaranteed functionality under POSIX, and indeed it was failing
on Linux under non-native filesystems (vfat, vboxfs).

This was fixed by write()ing one byte and then ftruncate()ing it back.

+++ libutils/file_lib.c
@@ -108,9 +108,9 @@ Writer *FileReadFromFd(int fd, size_t max_size, bool *truncated)
     }
 }
 
-int FullWrite(int desc, const char *ptr, size_t len)
+ssize_t FullWrite(int desc, const char *ptr, size_t len)
 {
-    int total_written = 0;
+    ssize_t total_written = 0;
 
     while (len > 0)
     {
@@ -134,13 +134,13 @@ int FullWrite(int desc, const char *ptr, size_t len)
     return total_written;
 }
 
-int FullRead(int fd, char *ptr, size_t len)
+ssize_t FullRead(int fd, char *ptr, size_t len)
 {
-    int total_read = 0;
+    ssize_t total_read = 0;
 
     while (len > 0)
     {
-        int bytes_read = read(fd, ptr, len);
+        ssize_t bytes_read = read(fd, ptr, len);
 
         if (bytes_read < 0)
         {
